# docs/openapi-agent-search.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anna Agent Search API",
    "version": "1.0.0",
    "description": "Hybrid (BM25 + semantic) search over the Vers3Dynamics engineering knowledge index, shaped for LLM tool calling: one POST, a flat citation-ready result list, and relevance scores normalized to 0-1. This spec is also served live at GET /api/v1/agent/openapi.json."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/v1/agent/search": {
      "post": {
        "operationId": "agentSearch",
        "summary": "Search the knowledge index (LLM tool contract)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked, filtered results.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request (bad type, missing query).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSearchError"
                }
              }
            }
          },
          "503": {
            "description": "Search backend unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSearchError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentSearchRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "query"
        ],
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "description": "Search string or academic query.",
            "examples": [
              "kalman filter divergence",
              "ESP32 DMA circular buffer"
            ]
          },
          "domain_filter": {
            "type": "string",
            "description": "Optional collection/topic/source filter, resolved in order: an explicit 'source:', 'kind:', 'category:' (alias 'topic:') prefix pins the facet; a bare document kind (paper, report, standard, repository, code, documentation, datasheet, library, other) filters by kind; a bare registered source name (arxiv, github, nasa, ...) filters by source; anything else filters by category. Unknown values match nothing rather than erroring.",
            "examples": [
              "arxiv",
              "paper",
              "category:cs.RO"
            ]
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 25,
            "default": 5,
            "description": "Maximum number of results to return. Out-of-range values are clamped, not rejected."
          },
          "min_score": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "default": 0.0,
            "description": "Relevance threshold on the normalized 0-1 score; results scoring below it are dropped. Out-of-range values are clamped."
          }
        }
      },
      "AgentSearchResult": {
        "type": "object",
        "required": [
          "id",
          "title",
          "authors",
          "content_chunk",
          "source_url",
          "relevance_score",
          "source",
          "kind",
          "published"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable document id, e.g. 'arxiv:8c4f01f9a3b2d715'."
          },
          "title": {
            "type": "string"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "content_chunk": {
            "type": "string",
            "description": "Query-relevant snippet: the retriever's matched passages when available, else the abstract, else the head of the body. Plain text, at most 1200 characters."
          },
          "source_url": {
            "type": "string",
            "description": "Canonical link for citation (falls back to the PDF link; empty string when the source has neither)."
          },
          "relevance_score": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "description": "Fused BM25+semantic rank score normalized to 0-1; 1.0 means top-ranked by every retriever that ran."
          },
          "source": {
            "type": "string",
            "description": "Ingestion source, e.g. 'arxiv', 'github', 'nasa'."
          },
          "kind": {
            "type": "string",
            "description": "Document kind, e.g. 'paper', 'report', 'code'."
          },
          "published": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO-8601 publication date, when known."
          }
        }
      },
      "AgentSearchResponse": {
        "type": "object",
        "required": [
          "status",
          "query",
          "total_results",
          "results"
        ],
        "properties": {
          "status": {
            "type": "string",
            "const": "success"
          },
          "query": {
            "type": "string",
            "description": "Echo of the query searched."
          },
          "total_results": {
            "type": "integer",
            "description": "Number of results returned after min_score filtering \u2014 always equals results.length, never a corpus-wide count."
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentSearchResult"
            }
          }
        }
      },
      "AgentSearchError": {
        "type": "object",
        "required": [
          "status",
          "error",
          "total_results",
          "results"
        ],
        "properties": {
          "status": {
            "type": "string",
            "const": "error"
          },
          "error": {
            "type": "string",
            "description": "Human-readable cause."
          },
          "total_results": {
            "type": "integer",
            "const": 0
          },
          "results": {
            "type": "array",
            "maxItems": 0,
            "items": {
              "$ref": "#/components/schemas/AgentSearchResult"
            }
          }
        }
      }
    }
  }
}